+++ phy/cluster/__init__.py
@@ -3,7 +3,6 @@
 
 """Automatic and manual clustering facilities."""
 
-from .algorithms import cluster, SpikeDetekt, KlustaKwik
 from .session import Session
 from .view_models import (BaseClusterViewModel,
                           HTMLClusterViewModel,
