+++ phy/cluster/algorithms/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Automatic clustering algorithms."""
