+++ phy/cluster/session.py
@@ -19,7 +19,7 @@ from ..io.base import BaseSession
 from ..io.kwik.model import KwikModel
 from ..io.kwik.store_items import create_store
 from .manual.gui import ClusterManualGUI
-from .algorithms import KlustaKwik, SpikeDetekt
+from .algorithms.klustakwik import KlustaKwik
 
 
 #------------------------------------------------------------------------------
