+++ phy/detect/tests/test_spikedetekt.py
@@ -17,11 +17,9 @@ from ...utils.datasets import _download_test_data
 from ...utils.logging import set_level
 from ...utils.testing import show_test
 from ...electrode.mea import load_probe
-from ...io.kwik import KwikModel
-from ...io.kwik.mock import create_mock_kwik
 from ...io.mock import artificial_traces
-from ..algorithms import (cluster, SpikeDetekt, _split_spikes,
-                          _concat, SpikeCounts)
+from ..spikedetekt import (SpikeDetekt, _split_spikes,
+                           _concat, SpikeCounts)
 
 
 #------------------------------------------------------------------------------
@@ -259,25 +257,3 @@ def test_spike_detect_real_data(tempdir, spikedetekt):
                     n_samples_per_spike=n_samples_w,
                     show=False)
     show_test(c)
-
-
-#------------------------------------------------------------------------------
-# Tests clustering
-#------------------------------------------------------------------------------
-
-def test_cluster(tempdir):
-    n_spikes = 100
-    filename = create_mock_kwik(tempdir,
-                                n_clusters=1,
-                                n_spikes=n_spikes,
-                                n_channels=8,
-                                n_features_per_channel=3,
-                                n_samples_traces=5000)
-    model = KwikModel(filename)
-
-    spike_clusters = cluster(model, num_starting_clusters=10)
-    assert len(spike_clusters) == n_spikes
-
-    spike_clusters = cluster(model, num_starting_clusters=10,
-                             spike_ids=range(100))
-    assert len(spike_clusters) == 100
